correzione dei bug per cui il programma funzionava solo con Android 13

+++ app/src/main/java/com/example/multipli/activityRecordAudio.kt
@@ -1,6 +1,8 @@
 package com.example.multipli
 
 
+import android.Manifest
+import android.annotation.SuppressLint
 import android.app.Notification
 import android.app.PendingIntent
 import android.app.Service
@@ -8,6 +10,7 @@ import android.content.Intent
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
+import android.content.pm.PackageManager
 import android.content.pm.ServiceInfo.FOREGROUND_SERVICE_TYPE_MICROPHONE
 import android.media.AudioManager
 import android.media.MediaRecorder
@@ -15,6 +18,10 @@ import android.os.Binder
 import android.os.Build
 import android.os.Environment
 import android.os.IBinder
+import androidx.annotation.RequiresApi
+import androidx.core.app.ActivityCompat
+import androidx.core.app.NotificationCompat
+import androidx.core.app.NotificationManagerCompat
 import java.io.File
 
  class activityRecordAudio: Service() {
@@ -36,14 +43,14 @@ import java.io.File
                      PendingIntent.FLAG_MUTABLE
                  )
              }
-         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+         if (Build.VERSION.SDK_INT > Build.VERSION_CODES.S) {
              val manager: NotificationManager =
                  getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
              val chan = NotificationChannel(
                  "Audio",
-                 "Audio", manager.importance
+                 "Audio", NotificationManager.IMPORTANCE_HIGH
              )
-             manager.createNotificationChannel(chan)
+            manager.createNotificationChannel(chan)
 
 
              val notification: Notification = Notification.Builder(this, "Audio")
@@ -58,8 +65,39 @@ import java.io.File
 
              startForeground(startId, notification, FOREGROUND_SERVICE_TYPE_MICROPHONE)
          }
+         else{
+             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                 val manager: NotificationManager =
+                     getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+                 val chan = NotificationChannel(
+                     "Audio",
+                     "Audio", NotificationManager.IMPORTANCE_HIGH
+                 )
+                 manager.createNotificationChannel(chan)
 
 
+                 val notification: Notification = Notification.Builder(this, "Audio")
+                     .setOngoing(true)
+                     .setContentTitle("Registrazione audio")
+                     .setContentText("sta registrando l'audio per poter campionare i valori di rumore")
+                     .setSmallIcon(R.drawable.ic_launcher_foreground)
+
+
+                     .build()
+                 if (ActivityCompat.checkSelfPermission(
+                         this,
+                         Manifest.permission.POST_NOTIFICATIONS
+                     ) == PackageManager.PERMISSION_GRANTED
+                 ) {
+                     with(NotificationManagerCompat.from(this)) {
+                         // notificationId is a unique int for each notification that you must define
+                         notify(13, notification)
+                     }
+                 }
+
+
+             }
+         }
 
          return START_STICKY
 
@@ -80,7 +118,10 @@ import java.io.File
 
 
 
-      fun audioOptionsFun(start:Boolean,context: Context)  {
+
+
+
+      fun audioOptionsFun(start:Boolean, context: Context)  {
          if (start){
 
          if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
@@ -108,20 +149,49 @@ import java.io.File
 
 
      }
+             else{
+             val audioDirect : String = Environment.getExternalStorageDirectory().path
+             val file = File(context.filesDir,"")
+             val file2 = File(file, "recording.mp3")
+             filetoDelate = file2
+                 audio = MediaRecorder()
+             audio.setAudioSource(MediaRecorder.AudioSource.MIC)
+             audio.setOutputFormat(MediaRecorder.OutputFormat.THREE_GPP)
+             audio.setAudioEncoder(MediaRecorder.AudioEncoder.AMR_NB)
+             AudioManager.PROPERTY_SUPPORT_AUDIO_SOURCE_UNPROCESSED
+             if (Build.VERSION.SDK_INT> Build.VERSION_CODES.Q){
+                 audio.setOutputFile(file2)
+             }
+             else {
+                 audio.setOutputFile("/dev/null")
+             }
+
+             audio.prepare()
+             ok= true
+             stop= true
+             audio.start()
+
+
+                 }
+
+
+             }
      }
 
-     }
+
      fun audioStop(){
          audio.stop()
-         filetoDelate.delete()
+         if (Build.VERSION.SDK_INT> Build.VERSION_CODES.Q) {
+             filetoDelate.delete()
+         }
          stop = false
      }
      fun startRecordingAudio(): Int {
            while (stop){
          amplitude = audio.maxAmplitude
 
-         //stopForeground(STOP_FOREGROUND_REMOVE)
-         //stopSelf()
+         stopForeground(STOP_FOREGROUND_REMOVE)
+         stopSelf()
          return amplitude
      }
          return 0
@@ -130,8 +200,8 @@ import java.io.File
 
      override fun onDestroy() {
          super.onDestroy()
-         stopForeground(STOP_FOREGROUND_REMOVE)
-         stopSelf()
+        stopForeground(STOP_FOREGROUND_REMOVE)
+         stopSelf()//
      }
      }
